api file remove 경고창 수정

+++ public/js/book/form.js
@@ -21,11 +21,11 @@ if(document.querySelector('#btRemoveCover'))
 if(document.querySelector('#btRemoveFile'))
     document.querySelector('#btRemoveFile').addEventListener('click',onRemoveFile);
 function onRemoveFile(e) {
-    console.log(this.dataset['idx'])
-    console.log(this.parentNode)
-    var idx =this.dataset['idx'];
-    var parent = this.parentNode;
-    axios.delete('/api/book/file/'+idx).then(onSucess).catch(onError);
+    if(confirm('파일을 삭제하시겠습니까?\n삭제하신 파일은 되돌릴 수 없습니다.')){
+        var idx =this.dataset['idx'];
+        var parent = this.parentNode;
+        axios.delete('/api/book/file/'+idx).then(onSucess).catch(onError);
+    }
 function onSucess(r){
     if(r.data.code === 200) parent.remove();
 }
